Add scheduler thread/runner init function
Add an init function callback that is invoked by each scheduler
running/thread at startup. This is needed to allow initialization to
happen that must take place on the thread itself. For example, the
custom backen initialization should be called from the scheduler
thread itself so that any CUDA set-device or other initialization can
happen on that thread, which is the same thread that will call into
the custom backend for execution.