Make CURL dependency PRIVATE in CMake
It seems like CURL should be a private dep(?).
Even though that means nothing for a static library, at least
it makes it clearer that the interface used by librequest(_static)
is not tied to CURL in any way.

Seems to me that CURL is not used in the public API and is
hidden in the C++ code. It should be the same for CMake.
It's the same for gRPC (although that seems to be linked
statically).

Signed-off-by: Andrei-Florin BENCSIK <andrei.bencsik@gmail.com>