Fix dynamic and sequence batcher to handle unloaded backend case
If the server is exiting or a backend is unloaded, it could be that a
scheduler thread for that backend is holding the last reference to the
backend. In this case the scheduler thread itself will destroy the
backend and thus the DynamicBatchScheduler or SequenceBatchScheduler
which owns the thread. So the thread needs to handle this case
correctly.