Description of correctly loaded models (#2341)
* Description of correctly loaded models

It was not clear that the output needed to display the models loaded in the table. I added this to the expected output and provided some explanations for the errors I ran into, which I presume could be common errors.

* Generalized model load in quickstart.md

Wrote the description of model load output in more general terms as requested by the review.

* Rephrase quickstart.md

Rephrase quickstart.md as suggested by pull request.