Fix nvtx annotations (#2799)
* NVTX: ignore `nvtxRangePushA` return value

For performance reasons, some tools will return a fixed value when
implementing the callback for `nvtxRangePushA`. Keeping track of the
nesting level for each thread requires accessing thread local data which
does not come for free.

For instance, Nsight Systems always returns `-1`. Following the previous
code logic, all ranges would be pushed but never popped.

Applications should not to detect failures in the tool via return codes
of NVTX API functions. It is intentionally something NVTX was designed to
prevent. We want the application to always take the same branches,
whether a tool is profiling the application or not.

This commit modifies the `NvtxRange` class to ignore the return value of
`nvtxRangePushA`.

Fixes #2798

* NVTX: overload `NvtxRange` for `const char *`

Overload the constructor to accept a C-style string. Previously, all
C-style strings were always implicitly converted to an `std::string`
which implies an unnecessary string copy.

Fixes #2798

* NVTX: run clang-format on `nvtx.h`

Fixes #2798

Co-authored-by: Antoine Froger <afroger@nvidia.com>