fix: Fix L0_implicit_state and it's variants (#7941)
Server waits for max_sequence_idle_microseconds before closing the request sequence and expects the client to start a NEW sequence for ANY subsequent request.
Test was failing because the max_sequence_idle_microseconds was set to low causing the server to believe the sequence is over and reject any further requests with SAME sequence_id without the START flag. Causing the test to fail.