Fixed fetch_models.sh to convert the inception TF model to an ONNX model
fetch_models.sh relied on a graphdef model being downloaded from Google
to showcase all the cloud examples. Since we are not supporting TF in 25.03
and later, we use the tf2onnx python module to convert this model to the
ONNX format. This should work as a drop in replacement for all tests,
documentation, and examples which rely on this model to function.