Dockerfile.QA includes updates from fetch_models.sh
Previously, Dockerfile.QA downloaded the TF inception models from Google
to include in some of the tests like L0_grpc and L0_http. This relied on
instructions from fetch_models.sh to download and install to the correct
directory. This commit updates this process to reflect the update to
fetch_models.sh to convert the graphdef model to an ONNX model.