scrubbed L0_warmup
This test relies on large models to show the effect of warming up
the model with data to reduce the latency on first inference.
However, a test which relies on large Tensorflow graphdef models
to show appreciable speedup doesnt compare the inference times
to ensure warmup works, it is purely demonstrative. Commenting
this section out because there are no suitible models, right
now, to drop in as replacements.

We do not lose code coverage with this decision since the
comparative times may be inversed and the test will stil pass.