scrubbed L0_backend_python
A number of python version tests rely on tensorflow being available
in pip for a python model to import. This has been switched with torch
as it can do the same function of testing the import and outputing
the version installed in the python model.

Starting in Ubuntu 24.04, only python 3.11+ is supported with the
default version of python being 3.12. All lower versions of python
being tested have been removed.