Part 3: scrubbed L0_infer
For this test all models are replicated amonst all the
backends. Interchanging Tensorflow models for ONNX models
was the easiest fix. For ensemble models, consistency should
be cared for when updating underlying models of the ensemble
such that calling test.sh:generate_model_repository() will
use the correct backends which are specified.'