docs: Reformat Model Repository Files
This change reformats a trio of model repository files in preparation for cleaning up and correcting documentation
related to model repository usage as well as the model un/load APIs.

No content changes are included.
Changes are focused on content flow and the correction of poor formatting which caused HTML rendering errors.